Moved back classes to original packages. Reduced the usage of private packages of dependencies. Made generator factory create code look like org.netbeans.modules.java.editor.codegen.ConstructorGenerator.java. Could not find a way to avoid the use of GeneratorUtils.
