+++ src/main/java/eu/hlavki/netbeans/svuid/MethodInfo.java
@@ -1,6 +1,6 @@
-package eu.hlavki.netbeans.svuid.model;
+package eu.hlavki.netbeans.svuid;
 
-import eu.hlavki.netbeans.svuid.model.ElementInfo;
+import eu.hlavki.netbeans.svuid.ElementInfo;
 import java.lang.reflect.Modifier;
 import java.util.Set;
 import javax.lang.model.element.Name;
