+++ src/main/java/eu/hlavki/netbeans/svuid/SerialVersionUidHintCustomizer.form
@@ -49,7 +49,7 @@
     <Component class="javax.swing.JCheckBox" name="warnForIncorrectValue">
       <Properties>
         <Property name="text" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="eu/hlavki/netbeans/svuid/ui/Bundle.properties" key="SerialVersionUidHintCustomizer.warnForIncorrectValue.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="eu/hlavki/netbeans/svuid/Bundle.properties" key="SerialVersionUidHintCustomizer.warnForIncorrectValue.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
       </Properties>
       <Events>
@@ -59,14 +59,14 @@
     <Component class="javax.swing.JLabel" name="ignoredValuesLabel">
       <Properties>
         <Property name="text" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="eu/hlavki/netbeans/svuid/ui/Bundle.properties" key="SerialVersionUidHintCustomizer.ignoredValuesLabel.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="eu/hlavki/netbeans/svuid/Bundle.properties" key="SerialVersionUidHintCustomizer.ignoredValuesLabel.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
       </Properties>
     </Component>
     <Component class="javax.swing.JTextField" name="ignoredValuesTextField">
       <Properties>
         <Property name="text" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="eu/hlavki/netbeans/svuid/ui/Bundle.properties" key="SerialVersionUidHintCustomizer.ignoredValuesTextField.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="eu/hlavki/netbeans/svuid/Bundle.properties" key="SerialVersionUidHintCustomizer.ignoredValuesTextField.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
       </Properties>
       <Events>
