+++ src/main/java/eu/hlavki/netbeans/svuid/SerialVersionUidHintCustomizer.java
@@ -1,4 +1,4 @@
-package eu.hlavki.netbeans.svuid.ui;
+package eu.hlavki.netbeans.svuid;
 
 import java.util.prefs.Preferences;
