+++ src/main/resources/eu/hlavki/netbeans/svuid/Bundle.properties
@@ -1,3 +1,17 @@
+# Copyright 2024 albertzb [albertzb42@gmail.com].
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 OpenIDE-Module-Display-Category=Java
 OpenIDE-Module-Long-Description=\
     SerialVersionUID Generator Plugin provides Java Hint and Code Generator to add default or generated serialVersionUID field of java.io.Serializable classes based on specification (<a href="http://java.sun.com/javase/6/docs/platform/serialization/spec/class.html">http://java.sun.com/javase/6/docs/platform/serialization/spec/class.html</a>).
@@ -11,5 +25,5 @@ HINT_SerialVersionUID_Generated=Add generated serialVersionUID
 LBL_SerialVersionGenerator_default=Add default serialVersionUID
 LBL_SerialVersionGenerator_generated=Add generated serialVersionUID
 SerialVersionUidHintCustomizer.warnForIncorrectValue.text=Warn for incorrect generated value
-SerialVersionUidHintCustomizer.ignoredValuesLabel.text=Ignored values:
 SerialVersionUidHintCustomizer.ignoredValuesTextField.text=
+SerialVersionUidHintCustomizer.ignoredValuesLabel.text=Ignored values:
